microservice_horizontal_serving_pattern/inference_server/requirements.txt
absl-py==1.1.0
alembic==1.8.0
anyio==3.6.1
astunparse==1.6.3
boto3==1.24.26
botocore==1.27.26
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.1.0
databricks-cli==0.17.0
docker==5.0.3
entrypoints==0.4
fastapi==0.78.0
Flask==2.1.2
flatbuffers==1.12
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.47.0
gunicorn==20.1.0
h11==0.13.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Mako==1.2.1
Markdown==3.3.7
MarkupSafe==2.1.1
mlflow==1.27.0
numpy==1.23.1
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.2
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
querystring-parser==1.2.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.6.0
scipy==1.8.1
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
starlette==0.19.1
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
typing_extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
websocket-client==1.3.3
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
